Redundant Convert.To analyzer (#18)
The analyzer finds all cases where the type of the result is the same as the type of the argument of the Convert.To* method. A code fix is offered to remove the call excluding cases when the result is not used. Then leaving the argument by itself could create an invalid code. The other option is to discard the argument by assigning it to _, but leaving it for a manual check seems a better option.